+++ development/circular-deps.jsonc
@@ -228,19 +228,6 @@
     "ui/pages/confirmations/confirmation/templates/error.js",
     "ui/pages/confirmations/confirmation/templates/index.js",
     "ui/selectors/index.js",
-    "ui/selectors/permissions.js",
-    "ui/selectors/selectors.js"
-  ],
-  [
-    "ui/components/component-library/index.ts",
-    "ui/components/component-library/picker-network/index.ts",
-    "ui/components/component-library/picker-network/picker-network.tsx",
-    "ui/components/multichain/avatar-group/avatar-group.tsx",
-    "ui/components/multichain/avatar-group/index.ts",
-    "ui/pages/confirmations/confirmation/ResultTemplate.ts",
-    "ui/pages/confirmations/confirmation/templates/error.js",
-    "ui/pages/confirmations/confirmation/templates/index.js",
-    "ui/selectors/index.js",
     "ui/selectors/selectors.js"
   ],
   [
