+++ ui/selectors/index.js
@@ -3,7 +3,6 @@ export * from './confirm-transaction';
 export * from './custom-gas';
 export * from './first-time-flow';
 export * from './metametrics';
-export * from './permissions';
 export * from './selectors';
 export * from './transactions';
 export * from './approvals';
